Camera system
Got some sort of camera system working. The object now appears at at a
random x/y coordinate depending on where it went off screen (so, if the
object leaves from the left side, it wiil re-appear from the right side
with a ranom y coordinate). I also changed the resolution of the game
so it looks and feels more open.

I also added a write_to_json function. It's also for the camera. I want
to save the players (and objects) x and y coords so when the player
enters an area where the objcet was for example, on the right side of
the screen, the object will not appear on the left. It will then read
from the json and render the square in the appropriate space. If that
makes sense.